Change reward calc library to 1.0.1 for now till we have 1.0.2 available

+++ app/src/main/java/com/bloxbean/cardano/yano/app/api/accounts/DebugSnapshotResource.java
@@ -74,7 +74,7 @@ public class DebugSnapshotResource {
             @JsonProperty("block_count") int blockCount,
             @JsonProperty("active_stake") BigInteger activeStake,
             @JsonProperty("delegator_count") int delegatorCount,
-            @JsonProperty("margin") double margin,
+            @JsonProperty("margin") BigDecimal margin,
             @JsonProperty("fixed_cost") BigInteger fixedCost,
             @JsonProperty("pledge") BigInteger pledge,
             @JsonProperty("reward_address") String rewardAddress,
